ImageStepper detects size from input file
Later, this will be input passed from the JavaScript and then stored in
the object as it does it's thing. I *did* just find a bug in the image
size, though, because I'm a dummy. It isn't 80x80. It's 80 by whatever-
is-necessary-to-make-a-3:2-ratio.

I figure I might as well take the input file so that it's always correct
for whatever sample image I'm using.