New image drawing function, iterates PPM to canvas
This function will draw the image pixel-by-pixel from the PPM data onto
the canvas. For some reason, Firefox 128 doesn't want to draw it...
except when artificially slowed down by setting a breakpoint between
each row. Firefox 130 doesn't seem to be affected, nor does Chromium
129.

This is hardly an optimal implementation, though, and I intend to switch
to using the ImageData object. Rust web_sys implements them, so it
should be easy enough for me to do.