Fill in the rest of the owl
I got it rendering!

I copied the code from my `rustpt` project -- an implementation of the
Raytracing In One Weekend book -- and fixed up the parts that fell
through.

Notably, the print statements and the random number generation. I'm no
longer printing strings to write as a PPM image at all, so that's gone
completely. Rust's `rand` crate doesn't support WASM on it's own, and
it looks like the only correct way to get random numbers is through the
browser. I decided to use `js-sys`'s Math::random function.

Now I just need to clean it all up. The `start` function should return
a Promise and run asynchronously. I need to emit status updates as the
rendering progresses -- probably via further manipulation of the DOM
from within Rust -- figure out how to cancel the render, and then
expose a bunch of controls to the user to customize the render. Probs
just the size, bounces, and samples-per-pixel. A scene editor is a bit
too involved.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch renderer-experimental
# Your branch is up to date with 'github/renderer-experimental'.
#
# Changes to be committed:
#	modified:   Cargo.toml
#	modified:   index.html
#	modified:   src/lib.rs
#	new file:   src/primitives.rs
#	new file:   src/renderer.rs
#	new file:   src/scene.rs
#
# Untracked files:
#	Cargo.lock
#